# Fix object and null possible type handling.