Remove ajv for multiple type handling.

The removal of ajv was not related with this fix but its need is showcased by the edited benchmark.

Before:
```
JSON.stringify obj x 1,624,624 ops/sec ±0.72% (91 runs sampled)
fast-json-stringify obj x 729,975 ops/sec ±0.96% (92 runs sampled)
fast-json-stringify-uglified obj x 730,695 ops/sec ±0.73% (93 runs sampled)
```
After:
```
JSON.stringify obj x 1,638,558 ops/sec ±1.08% (91 runs sampled)
fast-json-stringify obj x 6,023,692 ops/sec ±0.82% (92 runs sampled)
fast-json-stringify-uglified obj x 5,895,250 ops/sec ±0.91% (93 runs sampled)
```